+++ src/main/java/com/somemore/global/config/SecurityConfig.java
@@ -6,8 +6,8 @@ import com.somemore.global.auth.idpw.filter.IdPwAuthFilter;
 import com.somemore.global.auth.jwt.filter.JwtAuthFilter;
 import com.somemore.global.auth.jwt.filter.JwtExceptionFilter;
 import com.somemore.global.auth.jwt.usecase.GenerateTokensOnLoginUseCase;
-import com.somemore.global.auth.oauth.handler.failure.CustomOAuthFailureHandler;
-import com.somemore.global.auth.oauth.handler.success.CustomOAuthSuccessHandler;
+import com.somemore.global.auth.oauth.handler.CustomOAuthFailureHandler;
+import com.somemore.global.auth.oauth.handler.CustomOAuthSuccessHandler;
 import com.somemore.global.auth.oauth.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
